3. Тестовое задание. Hibernate. [#330581]

+++ src/main/java/ru/job4j/cars/model/PriceHistory.java
@@ -15,7 +15,7 @@ public class PriceHistory {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @EqualsAndHashCode.Include
     private int id;
-    private long before;
-    private long after;
+    private int before;
+    private int after;
     private LocalDateTime created;
 }
